Add the container logs feature
Selecting a container in the container view will
now present a logs view with the logs from that
container being updated.

There was a bit of an issue with the textual
layering...not sure I love how that works. If
two hidden widgets are on the same layer, the
widget rendered first is the only one you'll ever
see, given they're both 100% size.